sign up flow fixed

+++ src/app/common/ConformLeaveModal.js
@@ -36,15 +36,22 @@ class ConfirmLeaveModal extends BaseReactComponent {
         <Modal.Body>
           <div className="leave-modal-body">
             <p className="inter-display-medium f-s-20 lh-24 m-b-30 black-000">
-              Are you sure you want to leave ?{" "}
+              {this.props.customMessage
+                ? this.props.customMessage
+                : "Are you sure you want to leave ?"}
             </p>
             <div className="leave-modal-btn-section">
               <Button
                 className="secondary-btn m-r-24"
                 onClick={() => {
-                  resetUser();
-                  window.sessionStorage.setItem("refresh", false);
-                  this.props.history.push("/welcome");
+                  if (this.props.handleSignOutWelcome) {
+                    resetUser(true);
+                    this.props.handleSignOutWelcome();
+                  } else {
+                    resetUser();
+                    window.sessionStorage.setItem("refresh", false);
+                    this.props.history.push("/welcome");
+                  }
                 }}
               >
                 Yes
